security: split enterprise audit metadata
