+++ src/app/enterprise_audit.py
@@ -1,0 +1,62 @@
+import logging
+from datetime import datetime, timezone
+from typing import Any
+
+from app.enterprise_policy import enterprise_policy_version
+
+logger = logging.getLogger("enterprise_readiness")
+
+_SERVICE_NAME = "lotus-risk"
+_REDACT_FIELDS = {
+    "password",
+    "secret",
+    "token",
+    "authorization",
+    "ssn",
+    "account_number",
+    "client_email",
+}
+
+
+def redact_sensitive(value: Any) -> Any:
+    if isinstance(value, dict):
+        out: dict[str, Any] = {}
+        for key, item in value.items():
+            if key.lower() in _REDACT_FIELDS:
+                out[key] = "***REDACTED***"
+            else:
+                out[key] = redact_sensitive(item)
+        return out
+    if isinstance(value, list):
+        return [redact_sensitive(item) for item in value]
+    return value
+
+
+def emit_audit_event(
+    *,
+    action: str,
+    actor_id: str,
+    tenant_id: str,
+    role: str,
+    correlation_id: str | None,
+    metadata: dict[str, Any],
+) -> None:
+    logger.info(
+        "enterprise_audit_event",
+        extra={
+            "audit": {
+                "service": _SERVICE_NAME,
+                "action": action,
+                "actor_id": actor_id,
+                "tenant_id": tenant_id,
+                "role": role,
+                "correlation_id": correlation_id or "",
+                "timestamp_utc": datetime.now(timezone.utc).isoformat(),
+                "policy_version": enterprise_policy_version(),
+                "metadata": redact_sensitive(metadata),
+            }
+        },
+    )
+
+
+__all__ = ["emit_audit_event", "redact_sensitive"]
